LNTipServer/src/requirements.txt
astroid==2.0.1
beautifulsoup4==4.6.0
boto3==1.7.65
botocore==1.10.65
docutils==0.14
google==2.0.1
googleapis-common-protos==1.5.3
grpcio==1.13.0
grpcio-tools==1.13.0
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.6.0
pylint==2.0.1
python-dateutil==2.7.3
requests
s3transfer==0.1.13
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11